fix: move memory checks
