+++ walrus/methods/input/process_media.py
@@ -22,6 +22,7 @@ from imageio import imread
 from shared.image_tools import imresize
 
 from shared.database.user import UserbaseProject
+from shared.utils.memory_checks import check_and_wait_for_memory
 from shared.database.input import Input
 from shared.database.video.video import Video
 from shared.database.image import Image
@@ -139,24 +140,8 @@ def start_queue_check_loop(VIDEO_QUEUE, FRAME_QUEUE):
             add_deferred_items_time = 30  # reset
 
 
-def check_and_wait_for_memory(memory_limit_float = 75.0, check_interval = 5):
-    while True:
-        if is_memory_available(memory_limit_float = memory_limit_float):
-            return True
-        else:
-            logger.warn("No memory available, waiting. There is no harm if processing large amount please wait.")
-            time.sleep(check_interval)
 
 
-def is_memory_available(memory_limit_float = 75.0):
-    memory_percent = get_memory_percent()
-    if memory_percent is None: return True  # Don't stop if this check fails
-
-    if memory_percent > memory_limit_float:
-        logger.warn(f"[Memory] {memory_percent} % used is > {memory_limit_float} limit.")
-        return False
-
-    return True
 
 
 def check_if_add_items_to_queue(add_deferred_items_time, VIDEO_QUEUE, FRAME_QUEUE):
@@ -2489,15 +2474,3 @@ def clean_up_temp_dir(path):
         logger.error(f"shutil error {str(exc)}")
         pass
 
-
-def get_memory_percent():
-    import psutil
-
-    memory_percent = None
-    try:
-        psutil_memory_result = psutil.virtual_memory()
-        memory_percent = psutil_memory_result[2]
-    except Exception as e:
-        logger.warn(traceback.format_exc())
-
-    return memory_percent
